requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
certifi==2018.4.16
chardet==3.0.4
discord.py==0.16.12
idna==2.6
multidict==4.2.0
pkg-resources==0.0.0
pydiscourse==0.7.0
requests==2.18.4
urllib3==1.22
websockets==3.4
